pyproject.toml
[tool.poetry]
name = "object-translator-app"
version = "0.1.0"
description = ""
authors = ["Chad Gu <chad.m.gu@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
opencv-python = "^4.10.0.84"
streamlit = "^1.40.2"
logging = "^0.4.9.6"
aiohttp = "^3.11.10"
grpcio = "^1.68.1"
protobuf = "^5.29.1"
grpcio-tools = "^1.68.1"
googleapis-common-protos = "^1.66.0"
translators = "^5.9.3"
torchvision = "^0.20.1"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
